休出日数　休日日数　実装終了　未テスト

+++ nts.uk/uk.com/alarm.ctx/nts.uk.ctx.alarm.dom/src/main/java/nts/uk/ctx/alarm/dom/byemployee/check/checkers/prospect/HolidayWorkDaysProspector.java
@@ -35,8 +35,8 @@ public class HolidayWorkDaysProspector extends WorkDaysProspectorBase {
                     // 会社ID, 勤務種類, 労働制, 勤務種類のカウント表、出勤してるか
                     result.aggregate(item.getLaborSystem(), table, isAttendanceDay);
                     return result.getDays().v();
-                }).orElse((double)0);
-            }).orElse((double)0);
+                }).orElse(0.0);
+            }).orElse(0.0);
         });
     }
